Getiryemek_request/requirements.txt
requests>=2.28.2
pandas>=1.5.2
psycopg2>=2.9.5
python-dotenv>=0.21.0